move creation of the client instance outside of the benchmarked function
The previous attempt created the client in the decorator. As @basepi
pointed out, this still ends up being run in the callable that is
benchmarked, therefore falsifying any measurement by ~0.5ms.

The new approach only uses the decorator to signal that a client
instance should be provided. The client itself is created before calling
the function and passed in using a partial.